refactor: :recycle: move player classes into package
